Emit 'connection' before 'connect'
onConnect is attached to the 'connection' event, so it's possible that the user attach to the 'connect' event and get called previously to the connection was correctly done (if onConnect is used to finish the connection). Emitting the 'connection' event before the 'connect' one it ensures that onConnect is the first function being called and that the connection is correctly done.